Move CellCoords unit tests into the proper module
